Move download from homepage to File Explorer
