+++ libs/remix-ui/workspace/src/lib/providers/FileSystemProvider.tsx
@@ -5,7 +5,7 @@ import { Toaster } from '@remix-ui/toaster' // eslint-disable-line
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { FileSystemContext } from '../contexts'
 import { browserReducer, browserInitialState } from '../reducers/workspace'
-import { initWorkspace, fetchDirectory, removeInputField, deleteWorkspace, clearPopUp, publishToGist, createNewFile, setFocusElement, createNewFolder, deletePath, renamePath, copyFile, copyFolder, runScript, emitContextMenuEvent, handleClickFile, handleExpandPath, addInputField, createWorkspace, fetchWorkspaceDirectory, renameWorkspace, switchToWorkspace, uploadFile } from '../actions'
+import { initWorkspace, fetchDirectory, removeInputField, deleteWorkspace, clearPopUp, publishToGist, createNewFile, setFocusElement, createNewFolder, deletePath, renamePath, copyFile, copyFolder, runScript, emitContextMenuEvent, handleClickFile, handleExpandPath, addInputField, createWorkspace, fetchWorkspaceDirectory, renameWorkspace, switchToWorkspace, uploadFile, handleDownloadFiles } from '../actions'
 import { Modal, WorkspaceProps } from '../types'
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { Workspace } from '../remix-ui-workspace'
@@ -115,6 +115,10 @@ export const FileSystemProvider = (props: WorkspaceProps) => {
     await handleExpandPath(paths)
   }
 
+  const dispatchHandleDownloadFiles = async () => {
+    await handleDownloadFiles()
+  }
+
   useEffect(() => {
     dispatchInitWorkspace()
   }, [])
@@ -214,7 +218,8 @@ export const FileSystemProvider = (props: WorkspaceProps) => {
     dispatchRunScript,
     dispatchEmitContextMenuEvent,
     dispatchHandleClickFile,
-    dispatchHandleExpandPath
+    dispatchHandleExpandPath,
+    dispatchHandleDownloadFiles
   }
   return (
     <FileSystemContext.Provider value={value}>
